chore: move hoster model
