+++ app/src/main/java/eu/kanade/tachiyomi/ui/player/loader/EpisodeLoader.kt
@@ -1,9 +1,9 @@
 package eu.kanade.tachiyomi.ui.player.loader
 
+import aniyomix.source.Hoster
+import aniyomix.source.Hoster.Companion.toHosterList
 import eu.kanade.domain.items.episode.model.toSEpisode
 import eu.kanade.tachiyomi.animesource.AnimeSource
-import eu.kanade.tachiyomi.animesource.model.Hoster
-import eu.kanade.tachiyomi.animesource.model.Hoster.Companion.toHosterList
 import eu.kanade.tachiyomi.animesource.model.Video
 import eu.kanade.tachiyomi.animesource.online.AnimeHttpSource
 import eu.kanade.tachiyomi.data.download.anime.AnimeDownloadManager
