+++ app/src/main/java/eu/kanade/tachiyomi/ui/player/loader/HosterLoader.kt
@@ -1,7 +1,7 @@
 package eu.kanade.tachiyomi.ui.player.loader
 
+import aniyomix.source.Hoster
 import eu.kanade.tachiyomi.animesource.AnimeSource
-import eu.kanade.tachiyomi.animesource.model.Hoster
 import eu.kanade.tachiyomi.animesource.model.Video
 import eu.kanade.tachiyomi.animesource.online.AnimeHttpSource
 import eu.kanade.tachiyomi.ui.player.controls.components.sheets.HosterState
